+++ app/src/main/java/com/murraycole/ucrrunner/view/DAO/Stats.java
@@ -48,7 +48,7 @@ public class Stats {
         return duration;
     }
 
-    public void setDuration(int duration) {
+    public void setDuration(double duration) {
         this.duration = duration;
     }
